fix: types

+++ pims/pims/importer/importer.py
@@ -317,7 +317,7 @@ class FileImporter:
                 ImportEventType.FILE_ERROR,
                 self.upload_path, exeception=e
             )
-            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path.exists():
+            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path is not None and self.upload_path.exists():
                 self.upload_path.delete_upload_root()
             raise e
 
@@ -344,7 +344,7 @@ class FileImporter:
                         ImportEventType.ERROR_CONVERSION,
                         self.spatial_path
                     )
-                    if AUTO_DELETE_FAILED_UPLOAD and self.upload_path.exists():
+                    if AUTO_DELETE_FAILED_UPLOAD and self.upload_path is not None and self.upload_path.exists():
                         self.upload_path.delete_upload_root()
                     raise FormatConversionProblem()
             except Exception as e:
@@ -352,7 +352,7 @@ class FileImporter:
                     ImportEventType.ERROR_CONVERSION,
                     self.spatial_path, exception=e
                 )
-                if AUTO_DELETE_FAILED_UPLOAD and self.upload_path.exists():
+                if AUTO_DELETE_FAILED_UPLOAD and self.upload_path is not None and self.upload_path.exists():
                     self.upload_path.delete_upload_root()
                 raise FormatConversionProblem()
 
@@ -363,7 +363,7 @@ class FileImporter:
             spatial_format = SpatialReadableFormatFactory().match(self.spatial_path)
             if not spatial_format:
                 self.notify(ImportEventType.ERROR_NO_FORMAT, self.spatial_path)
-                if AUTO_DELETE_FAILED_UPLOAD and self.upload_path.exists():
+                if AUTO_DELETE_FAILED_UPLOAD and self.upload_path is not None and self.upload_path.exists():
                     self.upload_path.delete_upload_root()
                 raise NoMatchingFormatProblem(self.spatial_path)
             self.notify(
@@ -381,7 +381,7 @@ class FileImporter:
                     ImportEventType.ERROR_INTEGRITY_CHECK, self.spatial_path,
                     integrity_errors=errors
                 )
-                if AUTO_DELETE_FAILED_UPLOAD and self.upload_path.exists():
+                if AUTO_DELETE_FAILED_UPLOAD and self.upload_path is not None and self.upload_path.exists():
                     self.upload_path.delete_upload_root()
                 raise ImageParsingProblem(self.spatial)
             self.notify(ImportEventType.END_INTEGRITY_CHECK, self.spatial)
@@ -416,7 +416,7 @@ class FileImporter:
                 ImportEventType.ERROR_HISTOGRAM, self.histogram_path, image,
                 exception=e
             )
-            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path.exists():
+            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path is not None and self.upload_path.exists():
                 self.upload_path.delete_upload_root()
             raise FileErrorProblem(self.histogram_path)
 
@@ -432,7 +432,7 @@ class FileImporter:
             directory.mkdir()  # TODO: mode
         except (FileNotFoundError, FileExistsError, OSError) as e:
             self.notify(ImportEventType.FILE_ERROR, directory, exception=e)
-            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path.exists():
+            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path is not None and self.upload_path.exists():
                 self.upload_path.delete_upload_root()
             raise FileErrorProblem(directory)
 
@@ -445,7 +445,7 @@ class FileImporter:
                 shutil.move(origin, dest)
         except (FileNotFoundError, FileExistsError, OSError) as e:
             self.notify(ImportEventType.FILE_NOT_MOVED, origin, exception=e)
-            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path.exists():
+            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path is not None and self.upload_path.exists():
                 self.upload_path.delete_upload_root()
             raise FileErrorProblem(origin)
 
@@ -458,7 +458,7 @@ class FileImporter:
             )
         except (FileNotFoundError, FileExistsError, OSError) as e:
             self.notify(ImportEventType.FILE_ERROR, path, exception=e)
-            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path.exists():
+            if AUTO_DELETE_FAILED_UPLOAD and self.upload_path is not None and self.upload_path.exists():
                 self.upload_path.delete_upload_root()
             raise FileErrorProblem(path)
